+++ event-management-backend/src/main/java/ro/msg/event_management/controller/EventController.java
@@ -1,12 +1,9 @@
 package ro.msg.event_management.controller;
 
 import java.io.*;
-import java.text.SimpleDateFormat;
 import java.time.DateTimeException;
 import java.time.LocalDate;
 import java.time.LocalTime;
-import java.util.ArrayList;
-import java.util.Date;
 import java.util.List;
 import java.util.NoSuchElementException;
 import javax.servlet.http.HttpServletResponse;
@@ -63,9 +60,6 @@ import ro.msg.event_management.utils.SortCriteria;
 public class EventController {
 
     private final EventService eventService;
-    private final PictureService pictureService;
-    private final SublocationService sublocationService;
-    private final EventSublocationService eventSublocationService;
     private final Converter<Event, EventDto> convertToDto;
     private final Converter<EventDto, Event> convertToEntity;
     private final Converter<EventView, EventFilteringDto> converter;
@@ -323,308 +317,15 @@ public class EventController {
     @PostMapping(value = "/import")
     @PreAuthorize("hasAnyRole('ROLE_ADMIN')")
     public ResponseEntity exportEventsCsv(@RequestParam MultipartFile csv) throws IOException, OverlappingEventsException, ExceededCapacityException {
-        InputStream inputStream = csv.getInputStream();
-
-        var reader = new BufferedReader(new InputStreamReader(inputStream));
-
-        reader.readLine(); // skip header
-        String errorString = "";
-
-        while(reader.ready()){
-            String line = reader.readLine();
-            var fields = line.split(",");
-
-            var title = fields[0];
-            var subtitle = fields[1];
-            var status = fields[2];
-            var startDate = fields[3];
-            var endDate = fields[4];
-            var endHour = fields[5];
-            var startHour = fields[6];
-            var maxPeople = fields[7];
-            var description = fields[8];
-            var highlighted = fields[9];
-            var observations = fields[10];
-            var ticketsPerUser = fields[11];
-            var creator = fields[12];
-            var ticketInfo = fields[13];
-
-            var picturesUrls = fields[14];
-
-            var locationId = fields[17];
-
-            var event = new Event();
-            if(title.equals("")) {
-                errorString += "Skipped invalid row\n";
-                continue;
-            }
-            event.setTitle(title);
-
-            if(subtitle.equals("")) {
-                errorString += "Skipped invalid row\n";
-                continue;
-            }
-            event.setSubtitle(subtitle);
-
-            if(status.equals("")) {
-                errorString += "Skipped invalid row\n";
-                continue;
-            }
-            event.setStatus(Boolean.parseBoolean(status));
-
-            if(startDate.equals("")) {
-                errorString += "Skipped invalid row\n";
-                continue;
-            }
-            event.setStartDate(LocalDate.parse(startDate));
-
-            if(endDate.equals("")) {
-                errorString += "Skipped invalid row\n";
-                continue;
-            }
-            event.setEndDate(LocalDate.parse(endDate));
-
-            if(endHour.equals("")) {
-                errorString += "Skipped invalid row\n";
-                continue;
-            }
-            event.setEndHour(LocalTime.parse(endHour));
-
-            if(startHour.equals("")) {
-                errorString += "Skipped invalid row\n";
-                continue;
-            }
-            event.setStartHour(LocalTime.parse(startHour));
-
-            if(maxPeople.equals("")) {
-                errorString += "Skipped invalid row\n";
-                continue;
-            }
-            event.setMaxPeople(Integer.parseInt(maxPeople));
-
-            if(description.equals("")) {
-                errorString += "Skipped invalid row\n";
-                continue;
-            }
-            event.setDescription(description);
-
-            if(highlighted.equals("")) {
-                errorString += "Skipped invalid row\n";
-                continue;
-            }
-            event.setHighlighted(Boolean.parseBoolean(highlighted));
-
-            if(observations.equals("")) {
-                errorString += "Skipped invalid row\n";
-                continue;
-            }
-            event.setObservations(observations);
-
-            if(ticketsPerUser.equals("")) {
-                errorString += "Skipped invalid row\n";
-                continue;
-            }
-            event.setTicketsPerUser(Integer.parseInt(ticketsPerUser));
-
-            if(creator.equals("")) {
-                errorString += "Skipped invalid row\n";
-                continue;
-            }
-            event.setCreator(creator);
-
-            if(ticketInfo.equals("")) {
-                errorString += "Skipped invalid row\n";
-                continue;
-            }
-            event.setTicketInfo(ticketInfo);
-
-            if(locationId.equals("")){
-                errorString += "Skipped invalid row\n";
-                continue;
-            }
-
-            //add category
-            var categories = fields[16];
-            var ids4 = categories.split(";");
-            var list = new ArrayList<TicketCategory>();
-            for(String category: ids4){
-                var data = category.split("\\|");
-                var cat = new TicketCategory();
-                cat.setTitle(data[0]);
-                cat.setSubtitle(data[1]);
-                cat.setPrice(Float.parseFloat(data[2]));
-                cat.setDescription(data[3]);
-                cat.setTicketsPerCategory(Integer.parseInt(data[4]));
-                cat.setAvailable(Boolean.parseBoolean(data[5]));
-
-                list.add(cat);
-            }
-            event.setTicketCategories(list);
-
-            try {
-                Long.parseLong(locationId);
-            }catch (NumberFormatException e){
-                errorString += "Location id must be a number!\n";
-                continue;
-            }
-            event = eventService.saveEvent(event, Long.parseLong(locationId));
-
-            //add pictures to db
-            var urls = picturesUrls.split(";");
-            for(String url: urls){
-                var pic = new Picture();
-                pic.setEvent(event);
-                pic.setUrl(url);
-                pictureService.savePicture(pic);
-            }
-
-            //add eventsublocation to db
-            var sublocationIds = fields[15];
-            var ids2 = sublocationIds.split(";");
-            for(String id: ids2){
-                if(!id.equals("")) {
-                    try {
-                        var subLocation = sublocationService.findById(Long.parseLong(id));
-                        var eventSublocation = new EventSublocation();
-                        eventSublocation.setEvent(event);
-                        eventSublocation.setSublocation(subLocation);
-                        eventSublocationService.saveES(eventSublocation);
-                    }
-                    catch (Exception e){
-
-                    }
-                }
-            }
-        }
-
-        if(errorString.equals("")){
-            errorString = "Success";
-        }
-        return new ResponseEntity(errorString, HttpStatus.OK);
+        var result = eventService.saveCsv(csv);
+        return new ResponseEntity(result, HttpStatus.OK);
     }
 
     @GetMapping(value = "/export", produces = "text/csv")
     @PreAuthorize("hasAnyRole('ROLE_ADMIN')")
     public ResponseEntity<InputStreamResource> getEventsCsv() throws FileNotFoundException {
-        var events = eventService.findAll();
-
-        var headers = new ArrayList<String>();
-        headers.add("title");
-        headers.add("subtitle");
-        headers.add("status");
-        headers.add("startDate");
-        headers.add("endDate");
-        headers.add("endHour");
-        headers.add("startHour");
-        headers.add("maxPeople");
-        headers.add("description");
-        headers.add("highlighted");
-        headers.add("observations");
-        headers.add("ticketsPerUser");
-        headers.add("creator");
-        headers.add("ticketInfo");
-        headers.add("picturesUrls");
-        headers.add("sublocationIds");
-        headers.add("ticketCategoryTitle|ticketCategorySubtitle|ticketCategoryPrice|ticketCategoryDescription|ticketCategoryMax|ticketCategoryAvailable");
-
-        var date = new SimpleDateFormat("yyyy-MM-dd-hh-mm-ss").format(new Date());
-        var filename = "EventCSV" + date + ".csv";
-
-        try {
-            var writer = new FileWriter(filename);
-            for(String s:headers){
-                writer.write(s);
-                writer.write(',');
-            }
-
-            writer.write("\n");
-            for(Event event: events){
-                writer.write(event.getTitle());
-                writer.write(',');
-
-                writer.write(event.getSubtitle());
-                writer.write(',');
-
-                writer.write(String.valueOf(event.isStatus()));
-                writer.write(',');
-
-                writer.write(event.getStartDate().toString());
-                writer.write(',');
-
-                writer.write(event.getEndDate().toString());
-                writer.write(',');
-
-                writer.write(event.getEndHour().toString());
-                writer.write(',');
-
-                writer.write(event.getStartHour().toString());
-                writer.write(',');
-
-                writer.write(Integer.valueOf(event.getMaxPeople()).toString());
-                writer.write(',');
-
-                writer.write(event.getDescription());
-                writer.write(',');
-
-                writer.write(String.valueOf(event.isHighlighted()));
-                writer.write(',');
-
-                writer.write(event.getObservations());
-                writer.write(',');
-
-                writer.write(Integer.valueOf(event.getTicketsPerUser()).toString());
-                writer.write(',');
-
-                writer.write(event.getCreator());
-                writer.write(',');
-
-                writer.write(event.getTicketInfo());
-                writer.write(',');
-
-                var pictures = event.getPictures();
-                for(Picture picture: pictures){
-                    writer.write(String.valueOf(picture.getUrl()));
-                    writer.write(';');
-                }
-                writer.write(',');
-
-                var sublocations = event.getEventSublocations();
-                for(EventSublocation sublocation: sublocations){
-                    writer.write(String.valueOf(sublocation.getEventSublocationID().getSublocation()));
-                    writer.write(';');
-                }
-                writer.write(',');
-
-                var categories = event.getTicketCategories();
-                for(TicketCategory category: categories){
-                    writer.write(category.getTitle());
-                    writer.write('|');
-
-                    writer.write(category.getSubtitle());
-                    writer.write('|');
-
-                    writer.write(Float.valueOf(category.getPrice()).toString());
-                    writer.write('|');
-
-                    writer.write(category.getDescription());
-                    writer.write('|');
-
-                    writer.write(Integer.valueOf(category.getTicketsPerCategory()).toString());
-                    writer.write('|');
-
-                    writer.write(String.valueOf(category.isAvailable()));
-                    writer.write(';');
-                }
-                writer.write('\n');
-
-            }
-            writer.close();
-
-        } catch (IOException e){
-            e.printStackTrace();
-        }
-
-        return new ResponseEntity<>(new InputStreamResource(new FileInputStream(filename)), HttpStatus.OK);
+        var result = eventService.writeCsv();
+        return new ResponseEntity<>(result, HttpStatus.OK);
     }
 }
